% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amen-package.R
\docType{data}
\name{addhealthc3}
\alias{addhealthc3}
\title{AddHealth community 3 data}
\format{list}
\usage{
data(addhealthc3)
}
\description{
A valued sociomatrix (Y) and matrix of nodal attributes (X) for 
students in community 3 of the AddHealth study. 
\itemize{
\item Y: A sociomatrix in which the value of the edge corresponds to an ad-hoc measure of intensity of the relation. Note that students were only allowed to nominate up to 5 male friends and 5 female friends. 
\item X: Matrix of students attributes, including sex, race (1=white, 2=black, 3=hispanic, 4=asian, 5=mixed/other) and grade. 
} 
See \url{http://moreno.ss.uci.edu/data.html#adhealth} for more details.
}
\keyword{datasets}

