% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlines.R
\name{addlines}
\alias{addlines}
\title{Add lines}
\usage{
addlines(Y,X,col="lightblue",alength=0,...)
}
\arguments{
\item{Y}{a sociomatrix}

\item{X}{coordinates of nodes}

\item{col}{color of lines. Can be a vector of length equal to the number of edges to be drawn}

\item{alength}{length of arrows to be drawn}

\item{\ldots}{additional plotting parameters}
}
\description{
Add lines to a network plot
}
\examples{

data(addhealthc3) 
Y<-addhealthc3$Y
X<-xnet(Y) 
netplot(Y,X) 
addlines(Y,X,col=Y[Y!=0]) 

}
\author{
Peter Hoff
}

