% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayToList.R
\name{arrayToList}
\alias{arrayToList}
\title{Convert array to list.}
\usage{
arrayToList(arrayObj, actorList, sliceLabel)
}
\arguments{
\item{arrayObj}{3d array object}

\item{actorList}{list of actor names}

\item{sliceLabel}{labels for array slices}
}
\value{
array in list format
}
\description{
Convert array to list.
}
\author{
Shahryar Minhas
}

