% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comtrade}
\alias{comtrade}
\title{Comtrade data}
\format{A four-way array of yearly change in log trade between countries,
measured in 2000 US dollars across several commodity classes. The
four dimensions of the array index exporting nation, importing nation,
commidity and year, respectively.}
\source{
\url{http://comtrade.un.org/}
}
\description{
Summary of trade flows between countries over a ten year period.
}

