% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conflict90s}
\alias{conflict90s}
\title{Conflicts in the 90s}
\format{A list consisting of a sociomatrix \code{conflicts}, an array of
dyadic covariates \code{dyadvars} and nodal covariates \code{nodevars}.}
\source{
Michael Ward.
}
\description{
A relational dataset recording the total number of militarized disputes
or events between countries in the 1990s, along with nodal and dyadic
covariates.
}

