% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/DA.r
\name{DA}
\alias{DA}
\title{Function for classification of patients by chosen category via discriminant analysis}
\usage{
DA(data, variable, modelType = "MclustDA", modelNames = NULL)
}
\arguments{
\item{data}{Accepts a list with first row containing adult data and second row containing child data.}

\item{variable}{Accepts a variable upon which to attempt to classify the data}

\item{modelType}{Accepts the type of model to use. Options are MclustDA, which models the classes as a a mixture model of multiple normal distributions, or EDDA, which models each class as a single normal distribution. Defaults to MclustDA.}

\item{modelNames}{Accepts the name of the matrix constraint on the covariance matrix. See Mclust documentation for names of of constraints. Defaults to NULL.}
}
\description{
This function allows you to test classification of patients by some chosen category via discriminant analysis.
}
\examples{
DA()
}
\keyword{Discriminant}
\keyword{analysis}

