% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dutchcollege}
\alias{dutchcollege}
\title{Dutch college data}
\format{A list consisting of a socioarray \code{Y} and a matrix
\code{X}  of static nodal attributes. The relational
measurements range from -1 to 4, indicating the following:
\itemize{
\item -1 a troubled or negative relationship
\item  0 don't know
\item  1 neutral relationship
\item  2 friendly
\item  3 friendship
\item  4 best friends
}}
\source{
\url{http://moreno.ss.uci.edu/data.html#vdb}
}
\description{
Longitudinal relational measurements and nodal characteristics
of Dutch college students, described in
van de Bunt, van Duijn, and Snijders (1999).
}

