% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Fetch_Data.r
\name{Fetch_Data}
\alias{Fetch_Data}
\title{Function to fetch data}
\usage{
Fetch_Data(set_number = 1)
}
\arguments{
\item{set_number}{Specify whether to fetch the training set(1) or test set(2). Defaults to 1.}
}
\description{
This function fetches data contained in the Amen SPECT image datasets.
}
\examples{
Fetch_Data()
}
\keyword{Fetch}
\keyword{data}

