% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_EZ_cpp}
\alias{get_EZ_cpp}
\title{Linear combinations of submatrices of an array}
\usage{
Xbeta_cpp(X, beta)
}
\arguments{
\item{beta}{a p by 1 vector}

\item{ab}{a n by n matrix calculated by taking the outer 
product of the a and b random effect vectors}

\item{U}{an n by k matrix of multiplicative row effects}

\item{V}{an n by k matrix of multiplicative column effects}

\item{X}{a N length list filled with n by n by p arrays}
}
\value{
An n by n by N array
}
\description{
Computes a matrix of expected values based on an array X of predictors,
vector beta of regression coefficients, the outer product of the a and b
random effects, and the third order effects U and V.
}
\author{
Peter Hoff, Shahryar Minhas
}

