% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitObject.R
\name{getFitObject}
\alias{getFitObject}
\title{Get fitted object from MCMC results}
\usage{
getFitObject(APS, BPS, UVPS, YPS, BETA, VC, GOF, Xlist, actorByYr, startVals,
  symmetric, tryErrorChecks)
}
\arguments{
\item{APS}{summed additive sender random effects}

\item{BPS}{summed additive sender random effects}

\item{UVPS}{summed multiplicative random effects}

\item{YPS}{summed Y posterior predictive values}

\item{BETA}{Matrix of draws for regression coefficient estimates}

\item{VC}{Matrix of draws for variance estimates}

\item{GOF}{Matrix of draws for goodness of fit calculations}

\item{Xlist}{List based version of design array}

\item{actorByYr}{List of actors by time point}

\item{startVals}{startVals for future model run}

\item{symmetric}{logical indicating whether model is symmetric}
}
\value{
Fitted AME object
}
\description{
Get fitted object from MCMC results
}
\author{
Peter Hoff, Shahryar Minhas
}

