% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStartVals.R
\name{getStartVals}
\alias{getStartVals}
\title{Get fitted object from MCMC results}
\usage{
getStartVals(startVals, Y, model, xP, rvar, cvar, R)
}
\arguments{
\item{startVals}{List object that is null or contains 
starting values}

\item{Y}{dependent variable in array format}

\item{model}{character vector (e.g. 'bin', 'nrm') specifying
model type}

\item{xP}{number of exogenous covariates}

\item{rvar}{logical indicating whether to include sender random
effects}

\item{cvar}{logical indicating whether to include receiver random
effects}

\item{R}{Number of dimensions for multiplicative effects}
}
\value{
List of starting values for MCMC
}
\description{
Get fitted object from MCMC results
}
\author{
Peter Hoff, Shahryar Minhas
}

