% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofPlot.R
\name{gofPlot}
\alias{gofPlot}
\title{Plot results of goodness of fit tests}
\usage{
gofPlot(GOF, symmetric)
}
\arguments{
\item{GOF}{matrix of GOF values from ame_repL}

\item{symmetric}{logical: Is the sociomatrix symmetric by design?}

\item{rvar}{logical: fit row random effects (asymmetric case)?}

\item{cvar}{logical: fit column random effects (asymmetric case)?}
}
\value{
Plot for goodness of fit statistics
}
\description{
Plot results of goodness of fit tests
}
\author{
Shahryar Minhas
}

