% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofstats2.R
\name{gofstats2}
\alias{gofstats2}
\title{Goodness of fit statistics}
\usage{
gofstats2(Y)
}
\arguments{
\item{Y}{a relational data matrix}
}
\value{
a vector of gof statistics
}
\description{
Goodness of fit statistics evaluating degree
patterns
}
\examples{

data(YX_nrm) 

gofstats2(YX_nrm$Y) 


}
\author{
François Caron
}

