% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Hist_1D.R
\name{Hist_1D}
\alias{Hist_1D}
\title{Marginal Distribution Function}
\usage{
Hist_1D(data, print_adults = TRUE, print_children = TRUE)
}
\arguments{
\item{data}{Accepts a list with first row containing adult data and second row containing child data.}

\item{print_adults}{Determines whether histograms for adults are printed. Defaults to TRUE.}

\item{print_children}{Determines whether histograms for children are printed. Defaults to TRUE.}
}
\description{
Generates the marginal distribution of SPECT image brightness values in different regions of the brain
}
\examples{
Hist_1D()
}
\keyword{Create}
\keyword{distributions.}
\keyword{marginal}

