% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amen-package.R
\name{IR90s}
\alias{IR90s}
\title{International relations in the 90s}
\format{A list consisting of a socioarray \code{dyadvars} of 
dyadic variables and matrix  \code{nodevars} of nodal variables. 
The dyadic variables include
\itemize{
\item total number of conflicts;
\item exports (in billions of dollars);
\item distance (in thousands of kilometers);
\item number of shared IGOs (averages across the years);
\item polity interaction. 
} 
The nodal variables include
\itemize{
\item population (in millions); 
\item gdp (in billions of dollars); 
\item polity 
}}
\source{
Michael Ward.
}
\description{
A relational dataset recording 
a variety of nodal and dyadic variables on countries in the 1990s,
including information on conflicts, trade and other variables. 
Except for the conflict variable, the variables are averages
across the decade.
}

