% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToArray.R
\name{listToArray}
\alias{listToArray}
\title{Convert list to array}
\usage{
listToArray(actorList, actors, Y, Xdyad, Xrow, Xcol)
}
\arguments{
\item{actors}{vector of actors}

\item{Y}{dv in list format}

\item{Xdyad}{dyadic covariates in list format}

\item{Xrow}{sender covariates in list format}

\item{Xcol}{receiver covariates in list format}
}
\value{
transforms Y, Xdyad, Xrow, and Xcol to arrays
}
\description{
Convert list to array
}
\author{
Shahryar Minhas
}

