% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho_fc.R
\name{llsrmRho}
\alias{llsrmRho}
\title{SRM log likelihood evaluated on a grid of rho-values}
\usage{
llsrmRho(Y, Sab, rhos, s2 = 1)
}
\arguments{
\item{Y}{sociomatrix assumed to follow a mean-zero SRM distribution}

\item{Sab}{covariance matrix of additive effects}

\item{rhos}{vector of rho-values at which to calculate the log-likelihood}

\item{s2}{current value of s2}
}
\value{
a vector of log-likelihood values
}
\description{
Calculation of the SRM log-likelihood over a grid of rho-values
}
\author{
Peter Hoff
}
