% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Mean_Vectors.r
\name{Mean_Vectors}
\alias{Mean_Vectors}
\title{Function for generating mean vector of SPECT image brightness values in different regions of the brain}
\usage{
Mean_Vectors(data)
}
\arguments{
\item{data}{Accepts a list with first row containing adult data and second row containing child data.}
}
\description{
This function generates and plots a vector containing the mean SPECT image brightness value sin different regions of the brain.
}
\examples{
Mean_Vectors()
}
\keyword{Mean}

