% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramPlot.R
\name{paramPlot}
\alias{paramPlot}
\title{Plot parameter results from MCMC}
\usage{
paramPlot(mcmcData)
}
\arguments{
\item{mcmcData}{matrix of values from MCMC}
}
\value{
Trace plot and density distribution of 
estimated parameters
}
\description{
Plot parameter results from MCMC
}
\author{
Shahryar Minhas
}

