% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomputeX.R
\name{precomputeX}
\alias{precomputeX}
\title{Precomputation of design matrix quantities.}
\usage{
precomputeX(X)
}
\arguments{
\item{X}{a three way array, the design array for an AME model}
}
\value{
the same three-way array but with derived quantities 
as attributes.
}
\description{
Computation of a variety of quantities from the design array
to be used in MCMC model fitting algorithms.
}
\author{
Peter Hoff
}
