% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_ab_fc.R
\name{rbeta_ab_fc}
\alias{rbeta_ab_fc}
\title{Conditional simulation of additive effects and regression coefficients}
\usage{
rbeta_ab_fc(Z, Sab, rho, X, s2 = 1, iV0=NULL,m0=NULL)
}
\arguments{
\item{Z}{n X n (latent) normal relational matrix, assumed to follow a 
social relations regression model (with any multiplicative effects 
subtracted out).}

\item{Sab}{row and column covariance}

\item{rho}{dyadic correlation}

\item{X}{n x n x p covariate array}

\item{s2}{dyadic variance}

\item{iV0}{prior precision matrix for regression parameters}

\item{m0}{prior mean vector for regression parameters}
}
\value{
\item{beta}{regression coefficients} \item{a}{additive row effects}
\item{b}{additive column effects}
}
\description{
Simulates from the joint full conditional distribution of (beta,a,b)
in a social relations regression model
}
\author{
Peter Hoff
}
