% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_ab_rep_fc.R
\name{rbeta_ab_rep_fc}
\alias{rbeta_ab_rep_fc}
\title{Gibbs sampling of additive row and column effects and regression coefficient
with independent replicate relational data}
\usage{
rbeta_ab_rep_fc(Z.T,Sab,rho,X.T,s2=1)
}
\arguments{
\item{Z.T}{n x n x T array, with the third dimension for replicates.
Each slice of the array is a (latent) normal relational matrix, with
multiplicative effects subtracted out}

\item{Sab}{row and column covariance}

\item{rho}{dyadic correlation}

\item{X.T}{n x n x p x T covariate array}

\item{s2}{dyadic variance}
}
\value{
\item{beta}{regression coefficients} \item{a}{additive row effects}
\item{b}{additive column effects}
}
\description{
Simulates from the joint full conditional distribution of (a,b,beta),
assuming same additive row and column effects and regression coefficient
across replicates.
}
\author{
Peter Hoff, Yanjun He
}

