% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbeta_ab_rep_fc_cpp}
\alias{rbeta_ab_rep_fc_cpp}
\title{Gibbs sampling of additive row and column effects and regression coefficient
with independent replicate relational data}
\usage{
rbeta_ab_rep_cpp(Z.T,Sab,rho,X.T,s2=1)
}
\arguments{
\item{XrCube}{n x p x T row covariate array generated within ame_repL fn}

\item{XcCube}{n x p x T column covariate array generated within ame_repL fn}

\item{mXCube}{n^2 x p x T design array generated within ame_repL fn}

\item{mXtCube}{n^2 x p x T transposed design array generated within ame_repL fn}

\item{xxCube}{p x p x T regression sum of squares array generated within ame_repl fn}

\item{xxTCube}{p x p x T transposed regression sum of squares array generated 
within ame_repl fn}

\item{iSe2}{variance matrix}

\item{Sabs}{row and column covariance}

\item{k}{dimensions for row and column random effects}

\item{G}{eigenvalue calcs from Sab}

\item{e}{n x k gaussian error matrix}

\item{colE}{column sums of e}

\item{Z.T}{n x n x T array, with the third dimension for replicates.
Each slice of the array is a (latent) normal relational matrix, with
multiplicative effects subtracted out}
}
\value{
\item{beta}{regression coefficients} \item{a}{additive row effects}
\item{b}{additive column effects}
}
\description{
Simulates from the joint full conditional distribution of (a,b,beta),
assuming same additive row and column effects and regression coefficient
across replicates.
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

