% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_ab_repL_fc.R
\name{rbeta_ab_repL_fc}
\alias{rbeta_ab_repL_fc}
\title{Gibbs sampling of additive row and column effects and regression coefficient
with independent replicate relational data}
\usage{
rbeta_ab_repL_fc(Z.T,Sab,rho,X.T,s2=1)
}
\arguments{
\item{Sab}{row and column covariance}

\item{rho}{dyadic correlation}

\item{X.T}{T list of n x n x p covariate arrays}

\item{s2}{dyadic variance}

\item{T}{length list of Z.T n x n arrays, with the list used to contain replicates.
Each object in the list is a (latent) normal relational matrix, with
multiplicative effects subtracted out}
}
\value{
\item{beta}{regression coefficients} \item{a}{additive row effects}
\item{b}{additive column effects}
}
\description{
Simulates from the joint full conditional distribution of (a,b,beta),
assuming same additive row and column effects and regression coefficient
across replicates.
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

