% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbeta_rep_cpp}
\alias{rbeta_rep_cpp}
\title{Generates parameters to aid in Gibbs sampling of regression coefficient
with independent replicate relational data}
\usage{
rbeta_rep_cpp(ZT, to, td, Xr, Xc, mX, mXt, XX, XXt)
}
\arguments{
\item{Z.T}{n x n x T array, with the third dimension for replicates.
Each slice of the array is a (latent) normal relational matrix, with
multiplicative effects subtracted out}

\item{XrCube}{n x p x T row covariate array generated within ame_repL fn}

\item{XcCube}{n x p x T column covariate array generated within ame_repL fn}

\item{mXCube}{n^2 x p x T design array generated within ame_repL fn}

\item{mXtCube}{n^2 x p x T transposed design array generated within ame_repL fn}

\item{xxCube}{p x p x T regression sum of squares array generated within ame_repl fn}

\item{xxTCube}{p x p x T transposed regression sum of squares array generated 
within ame_repl fn}

\item{iSe2}{variance matrix}
}
\value{
list of parameters to feed into gibbs sampling of covariate,
a random effects, and b random effects
}
\description{
Generates parameters to aid in Gibbs sampling of regression coefficient
with independent replicate relational data
}
\author{
Shahryar Minhas
}

