% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbetaCPP_ab_rep_fc.R
\name{rbetaCPP_ab_rep_fc}
\alias{rbetaCPP_ab_rep_fc}
\title{Gibbs sampling of additive row and column effects and regression coefficient
with independent replicate relational data}
\usage{
rbetaCPP_ab_rep_fc(Z.T,Sab,rho,s2=1,...)
}
\arguments{
\item{Z.T}{n x n x T array, with the third dimension for replicates.
Each slice of the array is a (latent) normal relational matrix, with
multiplicative effects subtracted out}

\item{XrLong}{n x p x T row covariate array generated within ame_repL fn}

\item{XcLong}{n x p x T column covariate array generated within ame_repL fn}

\item{mXLong}{n^2 x p x T design array generated within ame_repL fn}

\item{mXtLong}{n^2 x p x T transposed design array generated within ame_repL fn}

\item{xxLong}{p x p x T regression sum of squares array generated within ame_repl fn}

\item{xxTLong}{p x p x T transposed regression sum of squares array generated 
within ame_repl fn}
}
\value{
\item{beta}{regression coefficients} \item{a}{additive row effects}
\item{b}{additive column effects}
}
\description{
Simulates from the joint full conditional distribution of (a,b,beta),
assuming same additive row and column effects and regression coefficient
across replicates.
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

