% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rrho_mh}
\alias{rrho_mh}
\title{Metropolis update for dyadic correlation}
\usage{
rrho_mh(E, rho, s2 = 1)
}
\arguments{
\item{E}{square residual relational matrix}

\item{rho}{current value of rho}

\item{s2}{current value of s2}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation
}
\examples{
## The function is currently defined as
function (E, rho, s2 = 1)
{
    EM <- cbind(E[upper.tri(E)], t(E)[upper.tri(E)])/sqrt(s2)
    emcp <- sum(EM[, 1] * EM[, 2])
    emss <- sum(EM^2)
    m <- nrow(EM)
    sr <- 2 * (1 - cor(EM)[1, 2]^2)/sqrt(m)
    rho1 <- rho + sr * qnorm(runif(1, pnorm((-1 - rho)/sr), pnorm((1 -
        rho)/sr)))
    lhr <- (-0.5 * (m * log(1 - rho1^2) + (emss - 2 * rho1 *
        emcp)/(1 - rho1^2))) - (-0.5 * (m * log(1 - rho^2) +
        (emss - 2 * rho * emcp)/(1 - rho^2))) + ((-0.5 * log(1 -
        rho1^2)) - (-0.5 * log(1 - rho^2)))
    if (log(runif(1)) < lhr) {
        rho <- rho1
    }
    rho
  }
}
\author{
Peter Hoff
}

