% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rrho_mh_rep}
\alias{rrho_mh_rep}
\title{Metropolis update for dyadic correlation with independent replicate data}
\usage{
rrho_mh_rep(E.T, rho, s2 = 1)
}
\arguments{
\item{E.T}{Array of square residual relational matrix series. The third
dimension of the array is for different replicates. Each slice of the array
according to the third dimension is a square residual relational matrix.}

\item{rho}{current value of rho}

\item{s2}{current value of s2}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation with independent replicate data.
}
\examples{
## The function is currently defined as
function(E.T,rho,s2=1)
{
  N<-dim(E.T)[3]
  EM<-NULL
  for (t in 1:N){
    E<-E.T[,,t]
    EM<-rbind(EM,cbind(E[upper.tri(E)],t(E)[upper.tri(E)] )/sqrt(s2))
  }

  emcp<-sum(EM[,1]*EM[,2])
  emss<-sum(EM^2)

  m<- nrow(EM)
  sr<- 2*(1-cor(EM)[1,2]^2)/sqrt(m)

  rho1<-rho+sr*qnorm( runif(1,pnorm( (-1-rho)/sr),pnorm( (1-rho)/sr)))

  lhr<-(-.5*(m*log(1-rho1^2)+(emss-2*rho1*emcp)/(1-rho1^2)))-
    (-.5*(m*log(1-rho^2 )+(emss-2*rho*emcp )/(1-rho^2 )))   +
    ( (-.5*log(1-rho1^2)) - (-.5*log(1-rho^2)) )

  if(log(runif(1))<lhr) { rho<-rho1 }
  rho
}
}
\author{
Peter Hoff, Yanjun He
}

