% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rrho_mh_rep_cpp}
\alias{rrho_mh_rep_cpp}
\title{Metropolis update for dyadic correlation with independent replicate data}
\usage{
rrho_mh_rep_cpp(E.T, rho, s2 = 1)
}
\arguments{
\item{rho}{current value of rho}

\item{s2}{current value of s2}

\item{E.T}{Array of square residual relational matrix series. The third
dimension of the array is for different replicates. Each slice of the array
according to the third dimension is a square residual relational matrix.}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation with independent replicate data.
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

