% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho_mh_repL.R
\name{rrho_mh_repL}
\alias{rrho_mh_repL}
\title{Metropolis update for dyadic correlation with independent replicate data}
\usage{
rrho_mh_repL(E.T, rho, s2 = 1)
}
\arguments{
\item{E.T}{List of square residual relational matrices. The list objects
contain different replicates. Each list object
is a square residual relational matrix.}

\item{rho}{current value of rho}

\item{s2}{current value of s2}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation with independent replicate data.
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

