% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rs2_rep_fc_cpp}
\alias{rs2_rep_fc_cpp}
\title{Gibbs update for dyadic variance with independent replicate relational data}
\usage{
rs2_rep_fc_cpp(E.T, rhoMat=solve(matrix(c(1,rho,rho,1),2,2)) )
}
\arguments{
\item{rhoMat}{inverted off-diagnoal matrix of rho}

\item{E.T}{Array of square residual relational matrix series. The third
dimension of the array is for different replicates. Each slice of the array
according to the third dimension is a square residual relational matrix}
}
\value{
a new value of s2
}
\description{
Gibbs update for dyadic variance with independent replicate relational data
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

