% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs2_repL_fc.R
\name{rs2_repL_fc}
\alias{rs2_repL_fc}
\title{Gibbs update for dyadic variance with independent replicate relational data}
\usage{
rs2_repL_fc(E.T, rho)
}
\arguments{
\item{E.T}{List of square residual relational matrices. The list
is for different replicates. Each object in the list
is a square residual relational matrix}

\item{rho}{current value of rho}
}
\value{
a new value of s2
}
\description{
Gibbs update for dyadic variance with independent replicate relational data
}
\author{
Peter Hoff, Yanjun He, Shahryar Minhas
}

