% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSab_fc.R
\name{rSab_fc}
\alias{rSab_fc}
\title{Gibbs update for additive effects covariance}
\usage{
rSab_fc(a,b,Sab0=NULL,eta0=NULL)
}
\arguments{
\item{a}{vector of row random effects}

\item{b}{vector of row random effects}

\item{Sab0}{prior (inverse) scale matrix for the prior distribution}

\item{eta0}{prior degrees of freedom for the prior distribution}
}
\description{
Gibbs update for additive effects covariance
}
\author{
Peter Hoff
}
