% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSuv_fc.R
\name{rSuv_fc}
\alias{rSuv_fc}
\title{Gibbs update for multiplicative effects covariance}
\usage{
rSuv_fc(U,V, Suv0=NULL,kappa0=NULL)
}
\arguments{
\item{U}{matrix of row random effects}

\item{V}{matrix of row random effects}

\item{Suv0}{prior (inverse) scale matrix for the prior distribution}

\item{kappa0}{prior degrees of freedom for the prior distribution}
}
\description{
Gibbs update for multiplicative effects covariance
}
\author{
Peter Hoff
}
