% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rZ_bin_fc_cpp}
\alias{rZ_bin_fc_cpp}
\title{Simulate Z based on a probit model}
\usage{
rZ_bin_fc_cpp(Z, EZ, rho, Y)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{Y}{square binary relational matrix}
}
\value{
a square matrix , the new value of Z
}
\description{
Simulates a random latent matrix Z given its expectation, dyadic correlation
and a binary relational matrix Y
}
\author{
Peter Hoff, Shahryar Minhas
}

