% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rZ_nrm_fc}
\alias{rZ_nrm_fc}
\title{Simulate missing values in a normal AME model}
\usage{
rZ_nrm_fc(Z, EZ, rho,s2, Y)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}

\item{Y}{square relational matrix}
}
\value{
a square matrix, equal to Y at non-missing values
}
\description{
Simulates missing values of a sociomatrix under a normal AME model
}
\examples{
## The function is currently defined as
rZ_nrm_fc<-function(Z,EZ,rho,s2,Y)
{
  ZS<-simY_nrm(EZ,rho,s2)
  diag(ZS)<-rnorm(nrow(Y),diag(EZ),sqrt(s2))
  Z[is.na(Y)]<-ZS[is.na(Y)]
  Z
}
}
\author{
Peter Hoff
}

