% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Scree_Plot.r
\name{Scree_Plot}
\alias{Scree_Plot}
\title{Function for generating Scree Plots of Amen Data}
\usage{
Scree_Plot(data)
}
\arguments{
\item{data}{Accepts a list with first row containing adult data and second row containing child data.}
}
\description{
This function generates Scree Plots of Amen Data.
}
\examples{
Scree_Plot
}
\keyword{Scree,}
\keyword{Variance}

