% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amen-package.R
\name{sheep}
\alias{sheep}
\title{Sheep dominance data}
\format{A list consisting of the following:
\itemize{
\item \code{dom}: a directed socioarray recording the number of 
dominance encounters. 
\item \code{age}: the age of each sheep in years. 
}}
\source{
\url{http://moreno.ss.uci.edu/data.html#sheep}
}
\description{
Number of dominance encounters between 28 female bighorn sheep. 
Cell (i,j) records the number of times sheep i dominated sheep j. 
From Hass (1991).
}

