% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_tob.R
\name{simY_tob}
\alias{simY_tob}
\title{Simulate a tobit relational matrix}
\usage{
simY_tob(EY, rho, s2)
}
\arguments{
\item{EY}{square matrix giving the expected value of the relational matrix}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}
}
\value{
a square matrix
}
\description{
Simulates a tobit relational matrix
}
\author{
Peter Hoff
}
