% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm2el.R
\name{sm2el}
\alias{sm2el}
\title{Sociomatrix to edgelist}
\usage{
sm2el(sm,directed=TRUE)
}
\arguments{
\item{sm}{a sociomatrix with possibly valued relations}

\item{directed}{if TRUE, only use the upper triangular part of the matrix to enumerate edges}
}
\value{
an edglist
}
\description{
Construction of an edgelist from a sociomatrix
}
\examples{

Y<-matrix(rpois(10*10,.5),10,10) ; diag(Y)<-NA
E<-sm2el(Y) 
el2sm(E) - Y 

}
\author{
Peter Hoff
}

