% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Variance_Heatmaps.r
\name{Variance_Heatmaps}
\alias{Variance_Heatmaps}
\title{Function for generated heatmap of the covariance matrix of the Amen data}
\usage{
Variance_Heatmaps(data)
}
\arguments{
\item{data}{Accepts a list with first row containing adult data and second row containing child data.}
}
\description{
This function generates a heatmap of the covariance matrix of the Amen data
}
\examples{
Variance-Heatmaps()
}

