% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zscores}
\alias{zscores}
\title{rank-based z-scores}
\usage{
zscores(y)
}
\arguments{
\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Computes the normal scores corresponding to the ranks of a data vector
}
\examples{
## The function is currently defined as
zscores<-function(y)
{
 qnorm( rank(y,na.last="keep",ties.method="average")/(1+sum(!is.na(y))) )
}
}
\author{
Peter Hoff
}

