% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amen-package.R
\name{comtrade}
\alias{comtrade}
\title{Comtrade data}
\format{A list consisting of a socioarray \code{Trade} and a vector 
\code{dollars2010} of inflation rates. The socioarray gives 
yearly trade volume (exports and imports) 
in dollars for 10 different commodity classes 
for eleven years between 229 countries. This gives a five-way 
array. The first index is the reporting country, so 
\code{Trade[i,j,t,k,1]} is what \code{i} reports for exports to 
\code{j}, but in general this is not the same as 
\code{Trade[j,i,t,k,2]}, what \code{j} reports as importing from \code{i}.}
\source{
\url{http://comtrade.un.org/}, \url{http://www.measuringworth.com/}
}
\description{
Eleven years of import and export data between 229 countries. 
The data use the SITC Rev. 1 commodity classification, aggregated at the 
first level (AG1).
}

