% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_array.R
\name{design_array}
\alias{design_array}
\title{Computes the design socioarray of covariate values}
\usage{
design_array(Xrow=NULL,Xcol=NULL,Xdyad=NULL,intercept=TRUE,n)
}
\arguments{
\item{Xrow}{an n x pr matrix of row covariates}

\item{Xcol}{an n x pc matrix of column covariates}

\item{Xdyad}{an n x n x pd array of dyadic covariates}

\item{intercept}{logical}

\item{n}{number of rows/columns}
}
\value{
an n x n x (pr+pc+pd+intercept) 3-way array
}
\description{
Computes the design socioarray of covariate values for an AME fit
}
\author{
Peter Hoff
}

