% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofstats.R
\name{gofstats}
\alias{gofstats}
\title{Goodness of fit statistics}
\usage{
gofstats(Y)
}
\arguments{
\item{Y}{a relational data matrix}
}
\value{
a vector of gof statistics
}
\description{
Goodness of fit statistics evaluating second and third-order dependence
patterns
}
\examples{

data(YX_nrm) 

gofstats(YX_nrm$Y) 


}
\author{
Peter Hoff
}

