% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mhalf}
\alias{mhalf}
\title{Symmetric square root of a matrix}
\usage{
mhalf(M)
}
\arguments{
\item{M}{a positive definite matrix}
}
\value{
a matrix \code{H} such that \code{H^2} equals \code{M}
}
\description{
Computes the symmetric square root of a positive definite matrix
}
\examples{
## The function is currently defined as
function (M)
{
    tmp <- eigen(M)
    tmp$vec \%*\% sqrt(diag(tmp$val, nrow = nrow(M))) \%*\% t(tmp$vec)
  }
}
\author{
Peter Hoff
}

