% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhalf.R
\name{mhalf}
\alias{mhalf}
\title{Symmetric square root of a matrix}
\usage{
mhalf(M)
}
\arguments{
\item{M}{a positive definite matrix}
}
\value{
a matrix \code{H} such that \code{H^2} equals \code{M}
}
\description{
Computes the symmetric square root of a positive definite matrix
}
\author{
Peter Hoff
}

