% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ame.R
\name{plot.ame}
\alias{plot.ame}
\title{Plot results of an AME object}
\usage{
\method{plot}{ame}(x, ...)
}
\arguments{
\item{x}{the result of fitting an AME model}

\item{...}{additional parameters (not used)}
}
\value{
a series of plots
}
\description{
A set of plots summarizing the MCMC routine for an AME fit, as well as some
posterior predictive checks.
}
\author{
Peter Hoff
}

