% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulation from a multivariate normal distribution}
\usage{
rmvnorm(n, mu, Sigma, Sigma.chol = chol(Sigma))
}
\arguments{
\item{n}{sample size}

\item{mu}{multivariate mean vector}

\item{Sigma}{covariance matrix}

\item{Sigma.chol}{Cholesky factorization of \code{Sigma}}
}
\value{
a matrix with \code{n} rows
}
\description{
Simulates a matrix where the rows are i.i.d. samples from a multivariate
normal distribution
}
\author{
Peter Hoff
}

