% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho_mh.R
\name{rrho_mh}
\alias{rrho_mh}
\title{Metropolis update for dyadic correlation}
\usage{
rrho_mh(Z, rho, s2 = 1,offset=0)
}
\arguments{
\item{Z}{n X n normal relational matrix}

\item{rho}{current value of rho}

\item{s2}{current value of s2}

\item{offset}{matrix of the same dimension as Z. It is assumed that 
Z-offset is equal to dyadic noise, so the offset should contain any 
additive and multiplicative effects (such as 
\code{Xbeta(X,beta+ U\%*\%t(V) +  outer(a,b,"+")  }   )}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation
}
\author{
Peter Hoff
}
