% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrho_mh.R
\name{rrho_mh}
\alias{rrho_mh}
\title{Metropolis update for dyadic correlation}
\usage{
rrho_mh(E, rho, s2 = 1)
}
\arguments{
\item{E}{square residual relational matrix}

\item{rho}{current value of rho}

\item{s2}{current value of s2}
}
\value{
a new value of rho
}
\description{
Metropolis update for dyadic correlation
}
\author{
Peter Hoff
}

