% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs2_fc.R
\name{rs2_fc}
\alias{rs2_fc}
\title{Gibbs update for dyadic variance}
\usage{
rs2_fc(E, rho)
}
\arguments{
\item{E}{square residual relational matrix}

\item{rho}{current value of rho}
}
\value{
a new value of s2
}
\description{
Gibbs update for dyadic variance
}
\author{
Peter Hoff
}

