% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs2_rep_fc.R
\name{rs2_rep_fc}
\alias{rs2_rep_fc}
\title{Gibbs update for dyadic variance with independent replicate relational data}
\usage{
rs2_rep_fc(E.T, rho)
}
\arguments{
\item{E.T}{Array of square residual relational matrix series. The third
dimension of the array is for different replicates. Each slice of the array
according to the third dimension is a square residual relational matrix}

\item{rho}{current value of rho}
}
\value{
a new value of s2
}
\description{
Gibbs update for dyadic variance with independent replicate relational data
}
\author{
Peter Hoff, Yanjun He
}

