% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rUV_fc.R
\name{rUV_fc}
\alias{rUV_fc}
\title{Gibbs sampling of U and V}
\usage{
rUV_fc(E, U, V, rho, s2 = 1, Psi0=NULL, kappa0=NULL )
}
\arguments{
\item{E}{square residual relational matrix}

\item{U}{current value of U}

\item{V}{current value of V}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}

\item{Psi0}{scale parameter for the prior for the covariance matrix of U,V}

\item{kappa0}{degrees of freedom parameter for the prior for the covariance matrix of U,V}
}
\value{
\item{U}{a new value of U} \item{V}{a new value of V}
}
\description{
A Gibbs sampler for updating the multiplicative effect matrices U and V
}
\author{
Peter Hoff
}
