% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rUV_fc.R
\name{rUV_fc}
\alias{rUV_fc}
\title{Gibbs sampling of U and V}
\usage{
rUV_fc(E, U, V, rho, s2 = 1, shrink=TRUE)
}
\arguments{
\item{E}{square residual relational matrix}

\item{U}{current value of U}

\item{V}{current value of V}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}

\item{shrink}{adaptively shrink the factors with a hierarchical prior}
}
\value{
\item{U}{a new value of U} \item{V}{a new value of V}
}
\description{
A Gibbs sampler for updating the multiplicative effect matrices U and V
}
\author{
Peter Hoff
}

