% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZ_nrm_fc.R
\name{rZ_nrm_fc}
\alias{rZ_nrm_fc}
\title{Simulate missing values in a normal AME model}
\usage{
rZ_nrm_fc(Z, EZ, rho,s2, Y)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}

\item{Y}{square relational matrix}
}
\value{
a square matrix, equal to Y at non-missing values
}
\description{
Simulates missing values of a sociomatrix under a normal AME model
}
\author{
Peter Hoff
}

