% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZ_ord_fc.R
\name{rZ_ord_fc}
\alias{rZ_ord_fc}
\title{Simulate Z given the partial ranks}
\usage{
rZ_ord_fc(Z, EZ, rho, Y)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{Y}{matrix of ordinal data}
}
\value{
a square matrix, the new value of Z
}
\description{
Simulates a random latent matrix Z given its expectation, dyadic correlation
and partial rank information provided by W
}
\author{
Peter Hoff
}

