% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rZ_rrl.R
\name{rZ_rrl_fc}
\alias{rZ_rrl_fc}
\title{Simulate Z given relative rank nomination data}
\usage{
rZ_rrl_fc(Z, EZ, rho, Y, YL)
}
\arguments{
\item{Z}{a square matrix, the current value of Z}

\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{Y}{square matrix of ranked nomination data}

\item{YL}{list of ranked individuals, from least to most preferred in each
row}
}
\value{
a square matrix, the new value of Z
}
\description{
Simulates a random latent matrix Z given its expectation, dyadic correlation
and relative rank nomination data
}
\details{
simulates Z under the constraints (1) Y[i,j]>Y[i,k] => Z[i,j]>Z[i,k]
}
\author{
Peter Hoff
}

