% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_bin.R
\name{simY_bin}
\alias{simY_bin}
\title{Simulate a network, i.e. a binary relational matrix}
\usage{
simY_bin(EZ, rho)
}
\arguments{
\item{EZ}{square matrix giving the expected value of the latent Z matrix}

\item{rho}{dyadic correlation}
}
\value{
a square binary matrix
}
\description{
Simulates a network, i.e. a binary relational matrix
}
\author{
Peter Hoff
}

