% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_frn.R
\name{simY_frn}
\alias{simY_frn}
\title{Simulate an relational matrix based on a fixed rank nomination scheme}
\usage{
simY_frn(EZ, rho, odmax, YO)
}
\arguments{
\item{EZ}{a square matrix giving the expected value of the latent Z matrix}

\item{rho}{dyadic correlation}

\item{odmax}{a scalar or vector giving the maximum number of nominations for
each node}

\item{YO}{a square matrix identifying where missing values should be
maintained}
}
\value{
a square matrix, where higher values represent stronger
relationships
}
\description{
Simulate an relational matrix based on a fixed rank nomination scheme
}
\author{
Peter Hoff
}

