% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simY_nrm}
\alias{simY_nrm}
\title{Simulate a normal relational matrix}
\usage{
simY_nrm(EY, rho, s2)
}
\arguments{
\item{EY}{square matrix giving the expected value of the relational matrix}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}
}
\value{
a square matrix
}
\description{
Simulates a normal relational matrix
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (EY, rho, s2)
{
    YS <- simZ(EY, rho, s2)
    diag(YS) <- NA
    YS
  }
}
\author{
Peter Hoff
}

