% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_nrm.R
\name{simY_nrm}
\alias{simY_nrm}
\title{Simulate a normal relational matrix}
\usage{
simY_nrm(EY, rho, s2)
}
\arguments{
\item{EY}{square matrix giving the expected value of the relational matrix}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}
}
\value{
a square matrix
}
\description{
Simulates a normal relational matrix
}
\author{
Peter Hoff
}

