% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_ord.R
\name{simY_ord}
\alias{simY_ord}
\title{Simulate an ordinal relational matrix}
\usage{
simY_ord(EZ, rho, Y)
}
\arguments{
\item{EZ}{square matrix giving the expected value of the latent Z matrix}

\item{rho}{scalar giving the within-dyad correlation}

\item{Y}{ordinal relational data matrix}
}
\value{
a square matrix
}
\description{
Simulates an ordinal relational matrix having a particular marginal
distribution
}
\author{
Peter Hoff
}

