% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simY_rrl.R
\name{simY_rrl}
\alias{simY_rrl}
\title{Simulate an relational matrix based on a relative rank nomination scheme}
\usage{
simY_rrl(EZ, rho, odobs, YO)
}
\arguments{
\item{EZ}{a square matrix giving the expected value of the latent Z matrix}

\item{rho}{dyadic correlation}

\item{odobs}{a scalar or vector giving the observed number of nominations
for each node}

\item{YO}{a square matrix identifying where missing values should be
maintained}
}
\value{
a square matrix, where higher values represent stronger
relationships
}
\description{
Simulate an relational matrix based on a relative rank nomination scheme
}
\author{
Peter Hoff
}

