% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simZ.R
\name{simZ}
\alias{simZ}
\title{Simulate Z given its expectation and covariance}
\usage{
simZ(EZ, rho, s2 = 1)
}
\arguments{
\item{EZ}{expected value of Z}

\item{rho}{dyadic correlation}

\item{s2}{dyadic variance}
}
\value{
a simulated value of Z
}
\description{
Simulate Z given its expectation and covariance
}
\author{
Peter Hoff
}

