% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ame.R
\name{summary.ame}
\alias{summary.ame}
\title{Summary of an AME object}
\usage{
\method{summary}{ame}(object, ...)
}
\arguments{
\item{object}{the result of fitting an AME model}

\item{...}{additional parameters (not used)}
}
\value{
a summary of parameter estimates and confidence intervals for an AME
fit
}
\description{
Summary method for an AME object
}
\author{
Peter Hoff
}

