% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xbeta.R
\name{Xbeta}
\alias{Xbeta}
\title{Linear combinations of submatrices of an array}
\usage{
Xbeta(X, beta)
}
\arguments{
\item{X}{an n by n by p array}

\item{beta}{a p by 1 vector}
}
\value{
An n by n matrix
}
\description{
Computes a matrix of expected values based on an array X of predictors and a
vector beta of regression coefficients.
}
\author{
Peter Hoff
}

