% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amen-package.R
\docType{data}
\name{YX_bin_long}
\alias{YX_bin_long}
\title{binary relational data and covariates}
\format{a list}
\usage{
data(YX_bin_long)
}
\description{
a synthetic dataset that includes longitudinal binary relational data
as well as information on covariates
}
\examples{

data(YX_bin_long)
gofstats(YX_bin_long$Y[,,1]) 

}
\keyword{datasets}

