% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{zscores}
\alias{zscores}
\title{rank-based z-scores}
\usage{
zscores(y,ties.method="average")
}
\arguments{
\item{y}{a numeric vector}

\item{ties.method}{method for dealing with ties}
}
\value{
a numeric vector
}
\description{
Computes the normal scores corresponding to the ranks of a data vector
}
\author{
Peter Hoff
}
